changed floats to decimals. would need a more elegant solution to push to graph ql

+++ src/GraphQL.Tests/Utilities/AstPrinterTests.cs
@@ -65,7 +65,7 @@ namespace GraphQL.Tests.Utilities
         [Fact]
         public void prints_float_value()
         {
-            double value = 3.33;
+            decimal value = 3.33M;
 
             var val = new FloatValue(value);
             var result = _printer.Visit(val);
